mm: Stop kswapd early when nothing's waiting for it to free pages
Keeping kswapd running when all the failed allocations that invoked it
are satisfied incurs a high overhead due to unnecessary page eviction
and writeback, as well as spurious VM pressure events to various
registered shrinkers. When kswapd doesn't need to work to make an
allocation succeed anymore, stop it prematurely to save resources.

Change-Id: Icdf8970b252901d0874f535d043c7b4ab09fe734
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>